+++ src/app/qearn/history/history.component.html
@@ -1,38 +1,69 @@
 <ng-container *transloco="let t">
-  <div class="content_container gap-4">
-    <table mat-table [dataSource]="dataSource" class="mat-elevation-z8">
-      <!-- Week-From Column -->
-      <ng-container matColumnDef="weekFrom">
-        <th mat-header-cell *matHeaderCellDef>Week From</th>
-        <td mat-cell *matCellDef="let element">{{ element.weekFrom }}</td>
-      </ng-container>
+  <div class="content_container stake-history">
+    <!-- Spinner -->
+    <div *ngIf="isLoading" class="loading-container">
+      <mat-spinner></mat-spinner>
+    </div>
 
-      <!-- Week-To Column -->
-      <ng-container matColumnDef="weekTo">
-        <th mat-header-cell *matHeaderCellDef>Week To</th>
-        <td mat-cell *matCellDef="let element">{{ element.weekTo }}</td>
-      </ng-container>
+    <!-- Table -->
+    <div *ngIf="!isLoading" class="table-responsive">
+      <table mat-table [dataSource]="dataSource" class="mat-elevation-z8" matSort>
+        <!-- Locked Epoch Column -->
+        <ng-container matColumnDef="lockedEpoch">
+          <th mat-header-cell *matHeaderCellDef mat-sort-header>Locked Epoch</th>
+          <td mat-cell *matCellDef="let element">{{ element.lockedEpoch }}</td>
+        </ng-container>
 
-      <!-- Early Unlock % Column -->
-      <ng-container matColumnDef="earlyUnlock">
-        <th mat-header-cell *matHeaderCellDef>Early Unlock %</th>
-        <td mat-cell *matCellDef="let element">{{ element.earlyUnlock }}</td>
-      </ng-container>
+        <!-- Locked Amount Column -->
+        <ng-container matColumnDef="lockedAmount">
+          <th mat-header-cell *matHeaderCellDef mat-sort-header>Locked Amount</th>
+          <td mat-cell *matCellDef="let element">{{ element.lockedAmount }}</td>
+        </ng-container>
 
-      <!-- Burn % Column -->
-      <ng-container matColumnDef="burn">
-        <th mat-header-cell *matHeaderCellDef>Burn %</th>
-        <td mat-cell *matCellDef="let element">{{ element.burn }}</td>
-      </ng-container>
+        <!-- Locked Weeks Column -->
+        <ng-container matColumnDef="lockedWeeks">
+          <th mat-header-cell *matHeaderCellDef mat-sort-header>Locked Weeks</th>
+          <td mat-cell *matCellDef="let element">{{ element.lockedWeeks }}</td>
+        </ng-container>
 
-      <!-- Boost % Column -->
-      <ng-container matColumnDef="boost">
-        <th mat-header-cell *matHeaderCellDef>Boost %</th>
-        <td mat-cell *matCellDef="let element">{{ element.boost }}</td>
-      </ng-container>
+        <!-- Total Locked Amount in Epoch Column -->
+        <ng-container matColumnDef="totalLockedAmountInEpoch">
+          <th mat-header-cell *matHeaderCellDef mat-sort-header>Total Locked Amount (this Epoch)</th>
+          <td mat-cell *matCellDef="let element">{{ element.totalLockedAmountInEpoch }}</td>
+        </ng-container>
 
-      <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
-      <tr mat-row *matRowDef="let row; columns: displayedColumns"></tr>
-    </table>
+        <!-- Current Bonus Amount in Epoch Column -->
+        <ng-container matColumnDef="currentBonusAmountInEpoch">
+          <th mat-header-cell *matHeaderCellDef mat-sort-header>Current Bonus Amount (this Epoch)</th>
+          <td mat-cell *matCellDef="let element">{{ element.currentBonusAmountInEpoch }}</td>
+        </ng-container>
+
+        <!-- Early Unlock % Column -->
+        <ng-container matColumnDef="earlyUnlockPercent">
+          <th mat-header-cell *matHeaderCellDef mat-sort-header>Early Locking Reward</th>
+          <td mat-cell *matCellDef="let element">{{ element.earlyUnlockPercent }}</td>
+        </ng-container>
+
+        <!-- Full Unlock % Column -->
+        <ng-container matColumnDef="fullUnlockPercent">
+          <th mat-header-cell *matHeaderCellDef mat-sort-header>Full Unlocking Reward</th>
+          <td mat-cell *matCellDef="let element">{{ element.fullUnlockPercent }}</td>
+        </ng-container>
+
+        <!-- Actions Column -->
+        <ng-container matColumnDef="actions">
+          <th mat-header-cell *matHeaderCellDef>Actions</th>
+          <td mat-cell *matCellDef="let element">
+            <div mat-raised-button class="unlock" color="primary" (click)="openEarlyUnlockModal(element)">
+              {{ t('qearn.Earned.unlockEarly') }}
+            </div>
+          </td>
+        </ng-container>
+
+        <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+        <tr mat-row *matRowDef="let row; columns: displayedColumns"></tr>
+      </table>
+    </div>
+    <mat-paginator [pageSizeOptions]="[5, 10, 20]" class="paginator" showFirstLastButtons></mat-paginator>
   </div>
 </ng-container>
